requirements.txt
Bootstrap_Flask==2.0.2
click~=8.1.3
Flask~=2.2.2
Flask_Login==0.6.2
Flask_Mail==0.9.1
Flask_SQLAlchemy==2.5.1
Flask_WTF==1.0.1
itsdangerous==2.0.1
python-dotenv~=0.20.0
scipy~=1.9.2
SQLAlchemy==1.4.40
Werkzeug~=2.2.2
WTForms==3.0.1
numpy~=1.23.4
email_validator==1.3.0
